backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic[email]==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# AI Libraries (MINIMAL)
openai==1.3.7

# Chat/WebSocket
websockets==11.0.3

# Additional utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4


